SIZE is short rather than int

+++ db/com/cdegroot/db/recman/FreePhysicalRowId.java
@@ -1,5 +1,5 @@
 /*
- *  $Id: FreePhysicalRowId.java,v 1.1 2000/04/03 12:13:48 cdegroot Exp $
+ *  $Id: FreePhysicalRowId.java,v 1.2 2000/04/11 06:07:07 boisvert Exp $
  *
  *  Physical row id's on free list
  *
@@ -7,8 +7,8 @@
  *  Copyright (C) 1999, 2000 Cees de Groot <cg@cdegroot.com>
  *
  *  This library is free software; you can redistribute it and/or
- *  modify it under the terms of the GNU Library General Public License 
- *  as published by the Free Software Foundation; either version 2 
+ *  modify it under the terms of the GNU Library General Public License
+ *  as published by the Free Software Foundation; either version 2
  *  of the License, or (at your option) any later version.
  *
  *  This library is distributed in the hope that it will be useful,
@@ -16,8 +16,8 @@
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  *  Library General Public License for more details.
  *
- *  You should have received a copy of the GNU Library General Public License 
- *  along with this library; if not, write to the Free Software Foundation, 
+ *  You should have received a copy of the GNU Library General Public License
+ *  along with this library; if not, write to the Free Software Foundation,
  *  Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA
  */
 package com.cdegroot.db.recman;
@@ -29,8 +29,8 @@ package com.cdegroot.db.recman;
 final class FreePhysicalRowId extends PhysicalRowId {
     // offsets
     private static final short O_SIZE = PhysicalRowId.SIZE; // int size
-    static final int SIZE = O_SIZE + Magic.SZ_INT;
-    
+    static final short SIZE = O_SIZE + Magic.SZ_INT;
+
     /**
      *  Constructs a physical rowid from the indicated data starting at
      *  the indicated position.
@@ -43,10 +43,10 @@ final class FreePhysicalRowId extends PhysicalRowId {
     int getSize() {
 	return block.readInt(pos + O_SIZE);
     }
-    
+
     /** Sets the size */
     void setSize(int value) {
 	block.writeInt(pos + O_SIZE, value);
     }
-    
+
 }
